forward: replace uniform disribution with normal
the uniform distribution was creating weights that
were too large and therefore were causing our softmax
function to have too large and too small numbers respective
to the mean. this produced an initial output layer result of
mostly zeros, which would cause training time to be too large.

a normal distribution allows us to stay closer to a mean of
input layer size and therefore create edge weights that are
easier to train.

uses the Kaiming Initialization technique